+++ include/JobFileB9J.h
@@ -20,8 +20,8 @@ class JobFileB9J : public JobFile {
 		std::string m_version;
 		//std::string m_xyPixel;
 		//std::string m_zLayer;
-		int m_xyPixel;
-		int m_zLayer;
+		double m_xyPixel;
+		double m_zLayer;
 		int32_t m_base;
 		int32_t m_filled;
 		std::string m_reserved1;
@@ -47,8 +47,8 @@ class JobFileB9J : public JobFile {
 		JobFileB9J(const char* filename);
 		~JobFileB9J();
 		void setScale(double scale);
-		int getXYResolution(){ return m_xyPixel; };
-		int getZResolution(){ return m_zLayer; };
+		double getXYResolution(){ return m_xyPixel; };
+		double getZResolution(){ return m_zLayer; };
 		
 
 	protected:
